% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\docType{methods}
\name{read-methods}
\alias{read-methods}
\alias{read}
\title{Methods for Function read}
\usage{
read(object)
}
\description{
Reads longitudinal data to create a SaemixData object (internal)
}
\keyword{internal}
