% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixModel.R
\docType{methods}
\name{[,SaemixModel-method}
\alias{[,SaemixModel-method}
\title{Get/set methods for SaemixModel object}
\usage{
\S4method{[}{SaemixModel}(x, i, j, drop)
}
\description{
Access slots of a SaemixModel using the object["slot"] format
}
\keyword{methods}
